Extract Spacer and LinkItem into separate files
