+++ packages/react/src/Pagination/LinkItem.tsx
@@ -1,0 +1,27 @@
+import { PaginationProps } from './Pagination'
+
+type LinkItemProps = Pick<PaginationProps, 'linkComponent' | 'linkTemplate'> & {
+  currentPage: PaginationProps['page']
+  pageNumber: number
+}
+
+export const LinkItem = ({ currentPage, linkComponent, linkTemplate, pageNumber }: LinkItemProps) => {
+  if (!linkComponent) return null
+
+  const Link = linkComponent
+
+  return (
+    <li>
+      <Link
+        aria-current={pageNumber === currentPage ? 'page' : undefined}
+        className="ams-pagination__link"
+        href={linkTemplate(pageNumber)}
+      >
+        <span className="ams-visually-hidden">
+          {pageNumber === currentPage ? `Pagina ${pageNumber}` : `Ga naar pagina ${pageNumber}`}
+        </span>
+        <span aria-hidden>{pageNumber}</span>
+      </Link>
+    </li>
+  )
+}
