+++ packages/react/src/Pagination/Spacer.tsx
@@ -1,0 +1,1 @@
+export const Spacer = () => <li aria-hidden>{'\u2026'}</li>
